--- Sparkle/Demo/DemoHelpers.swift
@@ -1,149 +1,0 @@
-//
-//  Copyright © FINN.no AS, Inc. All rights reserved.
-//
-import UIKit
-import SparkleCommon
-
-enum TabletDisplayMode {
-    case master
-    case detail
-    case fullscreen
-}
-
-public struct SparkleSection {
-    var title: String
-    var items: [SparkleItem]
-
-    public init(title: String, items: [SparkleItem]) {
-        self.title = title
-        self.items = items
-    }
-}
-
-public struct SparkleItem {
-    var title: String
-    var viewController: UIViewController
-
-    public init(title: String, viewController: UIViewController) {
-        self.title = title
-        self.viewController = viewController
-    }
-}
-//
-//enum Sections: String, CaseIterable {
-//    case demo
-//
-//    static var items: [Sections] {
-//        return allCases
-//    }
-//
-//    var numberOfItems: Int {
-//        return 0
-//    }
-//
-//    static func title(for section: Int) -> String {
-//        let section = Sections.items[section]
-//        let rawClassName = section.rawValue
-//        return rawClassName
-//    }
-//
-//    static func formattedNames(for section: Int) -> [String] {
-//        let section = Sections.items[section]
-//        let names: [String]
-//        switch section {
-//        case .demo:
-//            names = [""]
-//        }
-//        return names
-//    }
-//
-//    static func formattedName(for indexPath: IndexPath) -> String {
-//        let section = Sections.items[indexPath.section]
-//        var rawClassName: String
-//        switch section {
-//        case .demo:
-//            rawClassName = ""
-//        }
-//        return rawClassName.capitalizingFirstLetter
-//    }
-//
-//    static func `for`(_ indexPath: IndexPath) -> Sections {
-//        return Sections.items[indexPath.section]
-//    }
-//
-//    // swiftlint:disable:next cyclomatic_complexity
-//    static func viewController(for indexPath: IndexPath) -> UIViewController? {
-//        guard let section = Sections.items[safe: indexPath.section] else {
-//            return nil
-//        }
-//        var viewController: UIViewController?
-//        switch section {
-//        case .demo:
-//            viewController = UIViewController()
-//        }
-//
-//        let sectionType = Sections.for(indexPath)
-//        switch UIDevice.current.userInterfaceIdiom {
-//        case .pad:
-//            switch sectionType.tabletDisplayMode {
-//            case .master:
-//                if let unwrappedViewController = viewController {
-//                    viewController = SplitViewController(masterViewController: unwrappedViewController)
-//                }
-//            case .detail:
-//                if let unwrappedViewController = viewController {
-//                    viewController = SplitViewController(detailViewController: unwrappedViewController)
-//                }
-//            default:
-//                break
-//            }
-//        default:
-//            break
-//        }
-//
-//        let containmentOptions = (viewController as? Containable)?.containmentOptions ?? .none
-//        let shouldIncludeNavigationController = containmentOptions.contains(.navigationController)
-//        if shouldIncludeNavigationController {
-//            if let unwrappedViewController = viewController {
-//                viewController = NavigationController(rootViewController: unwrappedViewController)
-//            }
-//        }
-//
-//        let shouldIncludeTabBarController = containmentOptions.contains(.tabBarController)
-//        if shouldIncludeTabBarController {
-//            let tabBarController = UITabBarController()
-//            if let unwrappedViewController = viewController {
-//                tabBarController.viewControllers = [unwrappedViewController]
-//                viewController = tabBarController
-//            }
-//        }
-//
-//        let shouldPresentInBottomSheet = containmentOptions.contains(.bottomSheet)
-//        if shouldPresentInBottomSheet {
-//            if let unwrappedViewController = viewController {
-//                let bottomSheet = BottomSheet(rootViewController: unwrappedViewController)
-//                viewController = bottomSheet
-//            }
-//        }
-//
-//        return viewController
-//    }
-//
-//    var tabletDisplayMode: TabletDisplayMode {
-//        switch self {
-//        case .demo:
-//            return .fullscreen
-//        }
-//    }
-//}
-
-extension Array {
-    /// Returns nil if index < count
-    subscript(safe index: Int) -> Element? {
-        return indices.contains(index) ? self[index] : .none
-    }
-}
-
-extension Foundation.Notification.Name {
-    static let didChangeUserInterfaceStyle = Foundation.Notification.Name("didChangeUserInterfaceStyle")
-}
