+++ Sparkle/Demo/NotificationNames.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+extension Foundation.Notification.Name {
+    static let didChangeUserInterfaceStyle = Foundation.Notification.Name("didChangeUserInterfaceStyle")
+}
