+++ Sparkle/Sparkle.h
@@ -1,19 +1,4 @@
-//
-//  Sparkle.h
-//  Sparkle
-//
-//  Created by Elvis Nuñez on 12/11/2019.
-//  Copyright © 2019 FINN AS. All rights reserved.
-//
+@import Foundation;
 
-#import <Foundation/Foundation.h>
-
-//! Project version number for Sparkle.
 FOUNDATION_EXPORT double SparkleVersionNumber;
-
-//! Project version string for Sparkle.
 FOUNDATION_EXPORT const unsigned char SparkleVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <Sparkle/PublicHeader.h>
-
-
